new changes

+++ backend/routes/listings.js
@@ -29,7 +29,7 @@ listingRouter.get("/", async (req, res) => {
 listingRouter.get("/:id", async (req, res) => {
     const { id } = req.params;
     try {
-        const listing = await Listings.findById(id)
+        const listing = await Listings.findById(id).populate("userId","username")
         return res.json({
             listing: listing
         })
@@ -40,14 +40,17 @@ listingRouter.get("/:id", async (req, res) => {
     }
 
 })
-listingRouter.post("/", InputValidation(ListingSchema), userAuth, async (req, res) => {
+listingRouter.post("/", userAuth, async (req, res) => {
     const { title, image, price, description, bedrooms, guests } = req.body;
     const userId = req.user;
-    const hostedById = await User.findById(userId)
-    const username = hostedById.username.toString()
     try {
+        if(!title || !image || !price || !description || !bedrooms || !guests ){
+            return res.status(404).json({
+                message: "All inputs required"
+            })
+        }
         const NewListing = await Listings.create({
-            title, image, price, description, bedrooms, guests, hostedBy: username
+            title, image, price, description, bedrooms, guests, userId:userId
         })
         if (NewListing) {
             return res.json({
@@ -65,28 +68,25 @@ listingRouter.post("/", InputValidation(ListingSchema), userAuth, async (req, re
         })
     }
 })
-listingRouter.put("/:id", InputValidation(ListingSchema), userAuth, async (req, res) => {
+listingRouter.put("/:id", userAuth, async (req, res) => {
     const { title, image, price, description, bedrooms, guests } = req.body;
     const { id } = req.params
     const userId = req.user;
     try {
-        const findUser = await User.findById(userId)
-        const userUsername = findUser.username
-        const findListing = await Listings.findById(id)
-        if (findListing.hostedBy === userUsername) {
-            const NewListing = await Listings.findByIdAndUpdate({ _id: id }, {
+        const findUser = await Listings.findById(id)
+        if(findUser.userId == userId){
+             const NewListing = await Listings.findByIdAndUpdate({ _id: id }, {
                 title: title, image: image, price: price, description: description, bedrooms: bedrooms, guests: guests
             })
             return res.json({
                 message: "Listing updated successfully",
                 NewListing: NewListing
             })
-
         }
         return res.status(404).json({
-            message: "You are not owner of this listing"
+            message: "You are not owner"
         })
-
+         
     } catch (error) {
         res.status(404).json({
             message: error.message
@@ -97,44 +97,49 @@ listingRouter.delete("/:id", userAuth, async (req, res) => {
     const { id } = req.params;
     const userId = req.user;
     try {
-
-        if (userId) {
-            const DeleteListing = await Listings.findByIdAndDelete(id)
-            if (DeleteListing) {
-                return res.json({
-                    message: "Listing Deleted"
-                })
-            } else {
-                return res.status(404).json({
-                    message: error.message
-                })
+        const findOwner = await Listings.findById(id)
+        const owner = findOwner.userId
+        if(owner == userId){
+            if (userId) {
+                const DeleteListing = await Listings.findByIdAndDelete(id)
+                if (DeleteListing) {
+                    return res.json({
+                        message: "Listing Deleted"
+                    })
+                } else {
+                    return res.status(404).json({
+                        message: error.message
+                    })
+                }
             }
         }
+        return res.status(404).json({
+            message: "Not owner"
+        })
     } catch (error) {
         res.status(404).json({
             message: error.message
         })
     }
 })
-listingRouter.post("/:id", InputValidation(ReviewSchema), userAuth, async (req, res) => {
+// review
+listingRouter.post("/:id", userAuth, async (req, res) => {
     const { id } = req.params;
     const userId = req.user
     const { rate, comment } = req.body;
     try {
-        const findUser = await User.findById(userId)
         const findListing = await Listings.findById(id)
-        const rateBy = findUser.username
         if (findListing) {
             findListing.review.push({
                 rate: rate,
                 comment: comment,
-                rateBy: rateBy
+                rateBy: userId
             })
             findListing.save()
         }
-        return res.json({
-            message: "Thanks for rate"
-        })
+            return res.json({
+                message: "Thanks for rate"
+            })
     } catch (error) {
         res.status(404).json({
             message: error.message
@@ -145,6 +150,7 @@ listingRouter.get("/ratings/:id", async (req, res) => {
     const { id } = req.params;
     try {
         const findReview = await Listings.findById(id)
+        console.log(findReview)
         if (findReview.review === "") {
             return res.json({
                 review: ""
@@ -163,26 +169,29 @@ listingRouter.get("/ratings/:id", async (req, res) => {
 listingRouter.delete("/:id/reviews/:reviewId", userAuth, async (req, res) => {
     const { id, reviewId } = req.params;
     const userId = req.user
+    console.log("userid", userId)
     try {
         const user = await User.findById(userId)
-        const username = user.username
         const review = await Listings.findById(id)
         const reviewOwner = review.review.map((item) => item.rateBy)
-
-        if (!reviewOwner === username) {
-            return res.status(404).json({
-                messge: "You are not owner of this rate"
-            })
-        }
-        await Listings.findByIdAndUpdate(
+        console.log(reviewOwner)
+        if (reviewOwner == userId) {
+            await Listings.findByIdAndUpdate(
             id,
             {
                 $pull: { review: { _id: reviewId } }
             }
         )
-        return res.json({
-            message: "Review deleted!"
-        })
+            return res.json({
+               message: "Review deleted!"
+            })
+        }else{
+            return res.json({
+                message: "Not own rating"
+            })
+        }
+        
+        
     } catch (error) {
         res.status(404).json({
             message: error.message
